requirements.txt
pyTelegramBotAPI
python-dotenv
reportlab
prometheus-client
sentry-sdk
flask
pytest
pytest-cov